[로피] 3, 4단계 쇼핑 장바구니 제출합니다.  (#38)
* refactor: product id 타입을 Long으로 변경

* feat: CartProduct 데이터 클래스와 UI모델 생성

* feat: 장바구니 페이지 레이아웃 추가

* feat: 데이터베이스에 체크 표시에 관한 컬럼 생성 및 데이터베이스에서 장바구니에 체크된 아이템을 가져오는 함수 구현

* feat: 장바구니에서 체크 선택 및 수량 선택하여 가격 나타내는 기능 추가

* feat: 데이터 바인딩 연결

* docs[README]: 기능 목록 작성

* feat: 상품 상세에서 장바구니 담기 버튼을 누르면 다이얼로그가 열린다.
- 수량 선택 기능
- 다이얼로그에서 담기 버튼을 누르면 장바구니에 담기는 기능

* feat: 상품 상세에서 마지막으로 본 상품을 볼 수 있다.

* feat: 상품 상세에서 마지막으로 본 상품을 누르면 마지막으로 본 상품 상세로 이동한다.

* feat: 마지막으로 본 상품을 눌러 들어가면 마지막으로 본 상품을 볼 수 없도록 구현

* refactor: CustomViewOnClickListener 함수 파라미터 추가

* refactor: 상품을 클릭하면 상품 상세로 가는 기능 databinding 연결하여 리팩토링

* feat: 상품 목록에서 +버튼을 클릭할 수 있다.

* feat: 상품 목록에서 + 버튼을 클릭하면 수량을 선택할 수 있는 뷰가 열린다.

* feat: 상품 목록에서 보여지는 수량은 장바구니에 담겨있는 수량이다.

* feat: 수량을 선택할 수 있는 뷰에서 +/- 버튼을 누르면 수량이 1증가/1감소한다.
- 수량을 선택할 수 있는 뷰에서 +/- 버튼을 누르면 수량이 장바구니에 저장된다.

* refactor: 장바구니에 있는 아이템마다 수량을 +/- 로 선택할 수 있는 기능 listener와 livedata 사용하여 리팩토링

* refactor: 장바구니에 체크된 아이템들의 수량에 맞는 총 가격이 보여지는 기능 리팩토링, LiveData연결 및 databinding 활용

* feat: 장바구니에 전체를 선택할 수 있는 체크박스가 있다.
- 전체 체크박스를 선택하면 장바구니에 담긴 아이템들의 체크박스가 모두 선택
- 해제하면 장바구니에 담긴 아이템들의 체크박스 모두 해제
- 장바구니에 들어가 있는 아이템 개수가 몇 개인지 보여준다

* feat: 툴바에도 장바구니에 들어가 있는 아이템 개수가 몇 개인지 보여주는 기능 구현

* feat: 마지막으로 본 상품 페이지에서 뒤로 가기를 하면 상품 목록으로 이동한다.

* refactor: 장바구니페이지에서 목록으로 넘어가면서 변화되는 값들 모두 반영하도록 수정

* refactor: 불필요한 코드 제거

* feat: ShoppingPresenterTest 추가

* feat: ProductDetailPresenter 테스트 추가

* refactor: Offset 추상화

* refactor: CartItemsUIModel 제거하고 장바구니에서 얻을 수 있는 가격 데이터베이스에서 가져오도록 수정

* fix: 장바구니 페이지에서 돌아올 때 수량이 반영되지 않는 오류 수정

* feat: CartPresenter 테스트 추가

* feat: MockWebServer 연결하여 Product 가져옴

* fix: product item에서 + 버튼을 누르면 수량 선택하는 레이아웃이 안 뜨는 버그 수정

* refactor: callback 넘겨받아서 처리

* refactor: 데이터베이스 수정

* refactor: livedata 삭제

* refactor: gson 라이브러리 활용

+++ app/src/main/java/woowacourse/shopping/database/recentProduct/RecentProductDatabase.kt
@@ -44,7 +44,7 @@ class RecentProductDatabase(context: Context) : RecentRepository {
         val products = mutableListOf<Product>()
         while (cursor.moveToNext()) {
             val id =
-                cursor.getInt(cursor.getColumnIndexOrThrow(TABLE_COLUMN_ID))
+                cursor.getLong(cursor.getColumnIndexOrThrow(TABLE_COLUMN_ID))
             val name =
                 cursor.getString(
                     cursor.getColumnIndexOrThrow(TABLE_COLUMN_NAME),
@@ -63,7 +63,7 @@ class RecentProductDatabase(context: Context) : RecentRepository {
         return products
     }
 
-    override fun findById(id: Int): Product? {
+    override fun findById(id: Long): Product? {
         val selection = "$TABLE_COLUMN_ID = ?"
         val selectionArgs = arrayOf(id.toString())
         val cursor = db.query(
@@ -80,7 +80,7 @@ class RecentProductDatabase(context: Context) : RecentRepository {
         if (cursor.count > 0) {
             cursor.moveToFirst()
             val id =
-                cursor.getInt(cursor.getColumnIndexOrThrow(TABLE_COLUMN_ID))
+                cursor.getLong(cursor.getColumnIndexOrThrow(TABLE_COLUMN_ID))
             val name =
                 cursor.getString(
                     cursor.getColumnIndexOrThrow(TABLE_COLUMN_NAME),
@@ -99,7 +99,7 @@ class RecentProductDatabase(context: Context) : RecentRepository {
         return product
     }
 
-    override fun delete(id: Int) {
+    override fun delete(id: Long) {
         val whereClause = "$TABLE_COLUMN_ID = ?"
         val whereArgs = arrayOf(id.toString())
         db.delete(
